backend/requirements.txt
fastapi
uvicorn[standard]
langchain
faiss-cpu
sentence-transformers
pymupdf
beautifulsoup4
requests
python-dotenv
python-multipart
pydantic

